test np.array containing np.int32 elements with np.int32 type

+++ numpy/lib/tests/test_type_check.py
@@ -23,7 +23,7 @@ class TestCommonType(object):
         af64 = np.array([[1, 2], [3, 4]], dtype=np.float64)
         acs = np.array([[1+5j, 2+6j], [3+7j, 4+8j]], dtype=np.csingle)
         acd = np.array([[1+5j, 2+6j], [3+7j, 4+8j]], dtype=np.cdouble)
-        assert_(common_type(ai32) == np.float64)
+        assert_(common_type(ai32) == np.int32)
         assert_(common_type(af16) == np.float16)
         assert_(common_type(af32) == np.float32)
         assert_(common_type(af64) == np.float64)
